Fix changing data files crash.
Changing data files caused the data to become invalid, but the code to refresh
the view was triggered before the code to actually reload the data. Reload the
view when we actually need it, which is hopefully after the data have been
reloaded. More robust solutions welcome.

Ensure that the view can notice the change of data files by setting the data
count to a different value each time.

This has the unfortunate side-effect that the Accounts Summary view is reset to
the top level when _any_ change is made. This could be fixed with a bit more
work, but is unlikely to be a major annoyance.

Signed-off-by: Bruce Duncan <bwduncan@gmail.com>