Really speed up the loading.
Most of the time was spent doing string concatenation in getFullName.
String concatenation in Java is _really_ _slow_! Optimise the method to exit if
it finds a parent account with a previously-constructed fullName. Further
optimise the method to use StringBuilder instead of lots of string
concatenation.

This _noticeably_ speeds up the load in my case. The real kick in the teeth
was that I ran all my tests with the "Show long account names" option turned
OFF!

Signed-off-by: Bruce Duncan <bwduncan@gmail.com>