Again, walk the account tree in order.
This time in getAccountBalanceWithChildren. This avoids counting balances
multiple times.

However, this patch changes which accounts are included in the filtered list
in a way which I probably don't understand. Someone should check this.

Also, reinstate loadAccountBalances. Moving the calculations from the drawing
phase to the loading phase keeps everything together and this seems to speed
up my tests.

Signed-off-by: Bruce Duncan <bwduncan@gmail.com>